初始提交

+++ dbdatabase.cpp
@@ -11,12 +11,12 @@ bool DBDatabase::initDatabase()
     db.setDatabaseName("D:\\QT_C++_Project\\IoT-Device-Monitoring-Simulation\\test.db"); //資料庫名稱
     if(db.open()) //檢查開啟資料庫
     {
-        qDebug()<<"資料庫打開";
+        //qDebug()<<"資料庫打開";
         return db.open();
     }
     else
     {
-        qDebug()<<"資料庫打開失敗："<<db.lastError().text();
+        //qDebug()<<"資料庫打開失敗："<<db.lastError().text();
         return false;
     }
 }
@@ -35,24 +35,28 @@ void DBDatabase::closeDatabase(){
 bool DBDatabase::addDatabase
 (QString name, QString temp, QString rh, QString state, QString date_time)
 {
-    QSqlQuery query; //新增
-    query.prepare("INSERT INTO QT5(NAME,TEMP,RH,STATE,DATE_TIME)"
-                  "VALUES(:NAME, :TEMP, :RH, :STATE, :DATE_TIME)");
-    query.bindValue(":NAME", name);
-    query.bindValue(":TEMP", temp);
-    query.bindValue(":RH", rh);
-    query.bindValue(":STATE", state);
-    query.bindValue(":DATE_TIME", date_time);
-    if(query.exec())
+    if(db.open())
     {
-        qDebug()<<"新增成功";
-        return true;
+        QSqlQuery query; //新增
+        query.prepare("INSERT INTO QT5(NAME,TEMP,RH,STATE,DATE_TIME)"
+                      "VALUES(:NAME, :TEMP, :RH, :STATE, :DATE_TIME)");
+        query.bindValue(":NAME", name);
+        query.bindValue(":TEMP", temp);
+        query.bindValue(":RH", rh);
+        query.bindValue(":STATE", state);
+        query.bindValue(":DATE_TIME", date_time);
+        if(query.exec())
+        {
+            //qDebug()<<"新增成功";
+            return true;
+        }
+        else
+        {
+            //qDebug()<<"新增失敗："<<query.lastError().text();
+            return false;
+        }
     }
-    else
-    {
-        qDebug()<<"新增失敗："<<query.lastError().text();
-        return false;
-    }
+    return false;
 }
 
 /**
@@ -61,22 +65,27 @@ bool DBDatabase::addDatabase
  */
 bool DBDatabase::selectDatabase(QString fieldStr)
 {
-    QSqlQuery query("SELECT *, NAME FROM QT5"); //查詢
-    int fieldNo = query.record().indexOf(fieldStr);
-    if(query.isActive())
+    if(db.open())
     {
-        while (query.next())
+        QSqlQuery query("SELECT *, NAME FROM QT5"); //查詢
+        int fieldNo = query.record().indexOf(fieldStr);
+        if(query.isActive())
         {
-            QString str  = query.value(fieldNo).toString();
-            qDebug() <<"資料內容：" << str;
+            while (query.next())
+            {
+                QString str  = query.value(fieldNo).toString();
+                //qDebug() <<"資料內容：" << str;
+            }
+            return true;
         }
-        return true;
+        else
+        {
+            //qDebug()<<"查詢失敗："<<query.lastError().text();
+            return false;
+        }
     }
-    else
-    {
-        qDebug()<<"查詢失敗："<<query.lastError().text();
-        return false;
-    }
+    return false;
+
 }
 
 /**
@@ -90,25 +99,29 @@ bool DBDatabase::selectDatabase(QString fieldStr)
 bool DBDatabase::updateDatabase
 (int ID,QString name, QString temp, QString rh, QString state, QString date_time)
 {
-    QSqlQuery query; //修改
-    query.prepare("UPDATE QT5 SET NAME=:NAME, TEMP=:TEMP, "
-                  "RH=:RH, STATE=:STATE DATE_TIME=:DATE_TIME WHERE ID=:ID");
-    query.bindValue(":ID", ID);
-    query.bindValue(":NAME", name);
-    query.bindValue(":TEMP", temp);
-    query.bindValue(":RH", rh);
-    query.bindValue(":STATE", state);
-    query.bindValue(":DATE_TIME", date_time);
-    if(query.exec())
+    if(db.open())
     {
-        qDebug()<<"修改成功";
-        return true;
+        QSqlQuery query; //修改
+        query.prepare("UPDATE QT5 SET NAME=:NAME, TEMP=:TEMP, "
+                      "RH=:RH, STATE=:STATE DATE_TIME=:DATE_TIME WHERE ID=:ID");
+        query.bindValue(":ID", ID);
+        query.bindValue(":NAME", name);
+        query.bindValue(":TEMP", temp);
+        query.bindValue(":RH", rh);
+        query.bindValue(":STATE", state);
+        query.bindValue(":DATE_TIME", date_time);
+        if(query.exec())
+        {
+            //qDebug()<<"修改成功";
+            return true;
+        }
+        else
+        {
+            //qDebug()<<"修改失敗："<<query.lastError().text();
+            return false;
+        }
     }
-    else
-    {
-        qDebug()<<"修改失敗："<<query.lastError().text();
-        return false;
-    }
+    return false;
 }
 
 /**
@@ -117,17 +130,21 @@ bool DBDatabase::updateDatabase
  */
 bool DBDatabase::delteDatabase(int ID)
 {
-    QSqlQuery query; //刪除
-    query.prepare("DELETE FROM QT5 WHERE ID = :ID");
-    query.bindValue(":ID", ID);
-    if(query.exec())
+    if(db.open())
     {
-        qDebug()<<"刪除成功";
-        return true;
+        QSqlQuery query; //刪除
+        query.prepare("DELETE FROM QT5 WHERE ID = :ID");
+        query.bindValue(":ID", ID);
+        if(query.exec())
+        {
+            //qDebug()<<"刪除成功";
+            return true;
+        }
+        else
+        {
+            //qDebug()<<"刪除失敗："<<query.lastError().text();
+            return false;
+        }
     }
-    else
-    {
-        qDebug()<<"刪除失敗："<<query.lastError().text();
-        return false;
-    }
+    return false;
 }
